Provide function to regenerate RawStore within session (#4)
There is currently no easy way to regenerate the session id within Sessioner
meaning it is hard to update the session ID early on in a request tree.

RegenerateSession will look in the request scope for the session, grab the session
release, regenerate and then flush the old session before updating the RawStore in the
session.

Signed-off-by: Andrew Thornton <art27@cantab.net>

Reviewed-on: https://gitea.com/go-chi/session/pulls/4