Prevent panics due to EOF in gob.Decode (#5)
If a session file has been truncated due to server disk errors then the FileReader may return an io.EOF error on Read().

This then percolates up to the session manager which will panic.

This PR instead suggests that we simply swallow this error and start off again with empty session data.

Signed-off-by: Andrew Thornton <art27@cantab.net>

Co-authored-by: techknowlogick <techknowlogick@gitea.io>
Reviewed-on: https://gitea.com/go-chi/session/pulls/5
Co-authored-by: Andrew Thornton <art27@cantab.net>
Co-committed-by: Andrew Thornton <art27@cantab.net>